+++ lib/gg_golden.dart
@@ -6,4 +6,5 @@
 
 library;
 
-export 'src/gg_golden.dart';
+export 'src/expect_golden.dart';
+export 'src/write_golden.dart';